[merge] Split some code into merge.util
